+++ frontend/src/components/header/Header.tsx
@@ -1,28 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import { useLocation } from 'react-router-dom';
 import { SiteLogo } from '../siteLogo';
 import { UserProfile } from '../userProfile';
-
-interface pageTitles {
-  [key: string] : string, 
-}
+import usePageName from '../../hooks/usePageName';
 
 const Header = () => {
-  const [currentPath, setCurrentPath] = useState('');
-  const location = useLocation();
+  const { pageName } = usePageName();
 
-  const pageTitleByPath: pageTitles = {
-    'main': '최근 문서 목록',
-    'private' : '내 문서함',
-    'shared' : '공유 문서함',
-  };
-
-  useEffect(() => {
-    const lastPath = location.pathname.split('/').slice(-1);
-    setCurrentPath(pageTitleByPath[lastPath[0]]);
-  }, [location]);
-  
   const profileProps = {
     profileImgURL: 'https://picsum.photos/200',
     userName: 'iyu88'
@@ -31,7 +15,7 @@ const Header = () => {
   return (
     <HeaderWrapper>
       <SiteLogo />
-      <PageTitle>{currentPath}</PageTitle>
+      <PageName>{pageName}</PageName>
       <UserProfile {...profileProps}/>
     </HeaderWrapper>
   );
@@ -45,7 +29,7 @@ const HeaderWrapper = styled.div`
   border-bottom: 1px solid #BBBBBB;
 `;
 
-const PageTitle = styled.span`
+const PageName = styled.span`
   font-size: 1rem;
   font-weight: 500;
 `;
